refactor: home OpenAI and serve helper tests next to their modules
Move openai.rs unit tests into openai/tests.rs and relocate the
model-free bind/limits/server-config tests into serve.rs. CLI parse
tests that need Cli stay in main.rs so they can share ENV_LOCK.

Constraint: no production behavior change
Rejected: also move env-mutating bind tests | they share ENV_LOCK with CLI env tests
Confidence: high
Scope-risk: narrow
